Move interfaces to ports.
